requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.6.2
argilla==1.3.2
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.3.0
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.25
click==8.1.3
clickhouse-connect==0.5.25
colorama==0.4.6
coloredlogs==15.0.1
confection==0.0.4
contourpy==1.1.0
coverage==7.2.5
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.7
Deprecated==1.2.13
deprecation==2.0.7
dill==0.3.6
diskcache==5.6.1
distlib==0.3.6
dnspython==2.3.0
docker==6.1.0
docstring-parser==0.15
duckdb==0.8.0
duckduckgo-search==2.9.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl#sha256=35365d49d389ecb19ba33119dca9ab9ff80b36fbeb063fa5d44008197fead8fa
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fake-useragent==1.1.3
fastapi==0.92.0
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.12.0
flake8==6.0.0
flatbuffers==23.5.9
fonttools==4.41.0
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-search-results==2.4.2
googleapis-common-protos==1.59.0
gpt-engineer==0.0.7
gptcache==0.1.26
gradio==3.38.0
gradio_client==0.2.10
gTTS==2.3.1
gunicorn==20.1.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.16.4
humanfriendly==10.0
identify==2.5.24
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.176
langcodes==3.3.0
langflow==0.0.76
linkify-it-py==2.0.2
llama-cpp-python==0.1.50
llvmlite==0.40.0
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
murmurhash==1.0.9
mypy==1.3.0
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
nodeenv==1.7.0
numba==0.57.0
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.14.1
openai==0.27.8
openai-whisper==20230314
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.8.11
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pathspec==0.11.1
pathy==0.10.1
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.5.0
playsound==1.2.2
pluggy==1.0.0
posthog==3.0.1
pre-commit==3.3.3
preshed==3.0.8
protobuf==4.22.4
psycopg2-binary==2.9.6
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pydantic==1.10.7
pydub==0.25.1
pyflakes==3.0.1
Pygments==2.15.1
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.9.0
pyrsistent==0.19.3
PySocks==1.7.1
pysrt==1.1.2
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.1
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
readability-lxml==0.8.1
redis==4.5.4
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.5
rsa==4.9
rudder-sdk-python==2.0.2
ruff==0.0.272
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
selenium==4.9.0
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.15
srsly==2.4.6
starlette==0.25.0
sympy==1.12
tenacity==8.2.2
termcolor==2.3.0
thinc==8.1.10
threadpoolctl==3.2.0
tiktoken==0.3.3
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.14.0
typer==0.9.0
types-PyYAML==6.0.12.10
typing-inspect==0.8.0
typing_extensions==4.5.0
uc-micro-py==1.0.2
unstructured==0.5.13
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.23.0
wasabi==0.10.1
watchfiles==0.19.0
webdriver-manager==3.8.6
websocket-client==1.5.1
websockets==11.0.3
whisper==1.1.10
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.2.0
XlsxWriter==3.1.1
yarl==1.9.2
zstandard==0.21.0
